+++ Nos/Service/Filter.swift
@@ -7,38 +7,7 @@
 
 import Foundation
 
-struct RelaySubscription: Identifiable {
-    
-    let filter: Filter
-    
-    /// The date this Filter was opened as a subscription on relays. Used to close stale subscriptions
-    var subscriptionStartDate: Date?
-    
-    /// The number of objects using this filter. This is incremented and decremented by the RelayService to determine
-    /// when a filter can be closed.
-    var referenceCount: Int = 0
-    
-    var id: String {
-        subscriptionID
-    }
-    
-    // For closing requests; not part of hash
-    var subscriptionID: String {
-        filter.id
-    }
-    
-    var isActive: Bool {
-        subscriptionStartDate != nil
-    }
-    
-    /// Returns true if this is a "one-time" filter, where we are only looking for a single event
-    var isOneTime: Bool {
-        filter.limit == 1
-    }
-}
-
-
-/// For REQ
+/// Describes a set of Nostr Events, usually so we can ask relay servers for them.
 struct Filter: Hashable, Identifiable {
     
     let authorKeys: [String]
